fix .d.ts type output
