+++ src/index.ts
@@ -17,8 +17,8 @@
 import RRule from './rrule'
 import RRuleSet from './rruleset'
 import RRuleStr from './rrulestr'
-export { Frequency, ByWeekday, WeekdayStr, Options } from './types'
-export { default as Weekday } from './weekday'
+export { Frequency, ByWeekday, Options } from './types'
+export { Weekday, WeekdayStr } from './weekday'
 export { RRuleStrOptions } from './rrulestr'
 
 // =============================================================================
