+++ src/nlp/totext.ts
@@ -1,7 +1,7 @@
 import ENGLISH, { Language } from './i18n'
 import RRule from '../index'
 import { Options, ByWeekday } from '../types'
-import Weekday from '../weekday'
+import { Weekday } from '../weekday'
 import { isArray, isNumber, isPresent, padStart } from '../helpers'
 
 // =============================================================================
