+++ src/optionstostring.ts
@@ -1,7 +1,7 @@
 import { Options } from './types'
 import RRule, { DEFAULT_OPTIONS } from './rrule'
 import { includes, isPresent, isArray, isNumber } from './helpers'
-import Weekday from './weekday'
+import { Weekday } from './weekday'
 import dateutil from './dateutil'
 
 export function optionsToString (options: Partial<Options>) {
