+++ src/parsestring.ts
@@ -1,6 +1,7 @@
-import { Options, Frequency, Days } from './types'
-import Weekday from './weekday'
+import { Options, Frequency } from './types'
+import { Weekday } from './weekday'
 import dateutil from './dateutil'
+import { Days } from './rrule'
 
 export function parseString (rfcString: string) {
   rfcString = rfcString.replace(/^\s+|\s+$/, '')
